Extracting geography endpoints from policy author into their own service.
