+++ container-images/mds-geography/Dockerfile
@@ -1,0 +1,12 @@
+FROM node:12.14.1-alpine
+
+
+RUN apk add --no-cache tini
+
+RUN mkdir /mds
+
+COPY dist/* /mds/
+
+WORKDIR /mds
+
+ENTRYPOINT ["/sbin/tini", "node", "--no-deprecation", "index.js"]
