+++ container-images/mds-geography/tsconfig.build.json
@@ -1,0 +1,17 @@
+{
+  "extends": "../../tsconfig.build.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "references": [
+    {
+      "path": "../../container-images/env-inject/tsconfig.build.json"
+    },
+    {
+      "path": "../../packages/mds-geography/tsconfig.build.json"
+    },
+    {
+      "path": "../../packages/mds-api-server/tsconfig.build.json"
+    }
+  ]
+}
