+++ container-images/mds-geography/webpack.config.js
@@ -1,0 +1,6 @@
+module.exports = (env, argv) => require('../webpack.config')({
+  env,
+  argv,
+  dirname: require('path').resolve(__dirname),
+  bundles: ['index']
+})
