+++ helm/mds/values.yaml
@@ -54,6 +54,11 @@ apis:
     pathPrefix: /event-processor
     version: latest
     migration: false
+  mds-geography:
+    enabled: true
+    pathPrefix: /geography
+    version: latest
+    migration: false
 cronjobs:
   mds-trip-processor:
     enabled: true
