+++ packages/mds-geography/README.md
@@ -1,0 +1,2 @@
+# mds-geography
+Read and write for MDS Geography objects
