+++ packages/mds-geography/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@mds-core/mds-geography",
+  "description": "MDS Geography API",
+  "version": "0.0.1",
+  "private": true,
+  "licence": "Apache-2.0",
+  "engineStrict": true,
+  "engines": {
+    "node": ">= 8.0.0",
+    "npm": ">= 6.8.0"
+  },
+  "dependencies": {
+    "@mds-core/mds-api-helpers": "0.1.26",
+    "@mds-core/mds-api-server": "0.1.26",
+    "@mds-core/mds-db": "0.1.26",
+    "@mds-core/mds-logger": "0.1.24",
+    "@mds-core/mds-providers": "0.1.26",
+    "@mds-core/mds-schema-validators": "0.1.2",
+    "@mds-core/mds-utils": "0.1.26",
+    "express": "4.17.1",
+    "sinon": "7.5.0",
+    "uuid": "3.3.3"
+  },
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "build": "tsc --build tsconfig.build.json",
+    "start": "PATH_PREFIX=/geography yarn watch server",
+    "test": "yarn test:prettier && yarn test:eslint && yarn test:unit",
+    "test:eslint": "eslint --fix --ignore-path ../../.gitignore '**/*.ts'",
+    "test:prettier": "prettier --write --check --ignore-path ../../.gitignore '**/*.ts'",
+    "test:unit": "PATH_PREFIX=/geography DOTENV_CONFIG_PATH=../../.env nyc --lines 85 ts-mocha --project ../../tsconfig.json",
+    "watch": "nodemon --watch '../../packages' --ext 'ts' --ignore '*.d.ts' --exec yarn watch:exec --",
+    "watch:exec": "yarn build && ts-node"
+  }
+}
